move Message struct and rename
